+++ calendar-functions.ps1
@@ -8,7 +8,7 @@ Function Get-Calendar {
     [Alias("cal")]
 
     Param(
-        [Parameter(Position = 1,ParameterSetName = "month")]
+        [Parameter(Position = 1, ParameterSetName = "month")]
         [ValidateNotNullorEmpty()]
         [ValidateScript( {
                 $names = Get-MonthsByCulture
@@ -33,6 +33,15 @@ Function Get-Calendar {
 
         [Parameter(Mandatory, HelpMessage = "Enter an ending date for the month like 2/1/2019", ParameterSetName = "span")]
         [ValidateNotNullOrEmpty()]
+        [ValidateScript({
+            if ($_ -ge $Start) {
+                $True
+            }
+            else {
+                Throw "The end date ($_) must be later than the start date ($start)"
+                $False
+            }
+        })]
         [DateTime]$End,
 
         [string[]]$HighlightDate = (Get-Date).date.toString()
@@ -166,7 +175,7 @@ Function Show-Calendar {
             })]
         [string]$Month = (Get-Date -format MMMM),
 
-        [Parameter(Position = 2,ParameterSetName = "month")]
+        [Parameter(Position = 2, ParameterSetName = "month")]
         [ValidatePattern('^\d{4}$')]
         [int]$Year = (Get-Date).Year,
 
@@ -175,6 +184,8 @@ Function Show-Calendar {
         [ValidateNotNullOrEmpty()]
         [consolecolor]$HighlightColor = "Green"
     )
+
+    Write-Verbose "Starting $($myinvocation.mycommand)"
 
     #add default values if not bound
     $params = "Month", "Year", "HighlightDate"
@@ -227,7 +238,232 @@ Function Show-Calendar {
             Write-host $line
         }
     } #foreach line in calarray
+
+    Write-Verbose "Ending $($myinvocation.mycommand)"
+
 } #end Show-Calendar
+
+#create a WPF-based calendar
+Function Show-GuiCalendar {
+    [cmdletbinding()]
+    [OutputType("None")]
+    [Alias("gcal")]
+
+    Param(
+        [Parameter(Position = 1, HelpMessage = "Enter the first month to display by date, like 1/1/2019.")]
+        [ValidateNotNullOrEmpty()]
+        [datetime]$Start = [datetime]::new([datetime]::now.year, [datetime]::now.month, 1),
+
+        [Parameter(Position = 2, HelpMessage = "Enter the last month to display by date, like 3/1/2019. You cannot display more than 3 months.")]
+        [ValidateNotNullOrEmpty()]
+        [ValidateScript({
+            if ($_ -ge $Start) {
+                $True
+            }
+            else {
+                Throw "The end date ($_) must be later than the start date ($start)"
+                $False
+            }
+        })]
+        [datetime]$End = [datetime]::new([datetime]::now.year, [datetime]::now.month, 1),
+
+        [Parameter(HelpMessage = "Enter an array of dates to highlight like 12/25/2019.")]
+        [datetime[]]$HighlightDate,
+
+        [Parameter(HelpMessage = "Select a font family for your calendar" )]
+        [ValidateSet("Segoi UI", "QuickType", "Tahoma", "Lucida Console", "Century Gothic")]
+        [string]$Font = "Segoi UI",
+
+
+        [Parameter(HelpMessage = "Select a font style for your calendar." )]
+        [ValidateSet("Normal", "Italic", "Oblique")]
+        [string]$FontStyle = "Normal",
+
+        [Parameter(HelpMessage = "Select a font weight for your calendar." )]
+        [ValidateSet("Normal", "DemiBold", "Light", "Bold")]
+        [string]$FontWeight = "Normal"
+    )
+
+    Write-Verbose "Starting $($myinvocation.mycommand)"
+
+    if ($psedition -eq 'Core') {
+        Write-Warning "This function requires Windows PowerShell."
+        #bail out
+        Return
+    }
+
+    $months = do {
+        $start
+        $start = $start.AddMonths(1)
+    } while ($start -le $end)
+
+    if ($months.count -gt 3) {
+        Write-Warning "You can't display more than 3 months at a time with this command."
+        #bail out
+        Return
+    }
+
+    #add the necessary type library and bail out if there are errors
+
+    Try {
+        Add-Type -AssemblyName PresentationFramework -ErrorAction Stop
+        Add-Type –assemblyName PresentationCore -ErrorAction Stop
+        Add-Type –assemblyName WindowsBase -ErrorAction Stop
+        }
+    Catch {
+            Write-Warning "Failed to load a required type library. $($_.exception.message)"
+            #bail out
+            Return
+        }
+
+    $myParams=@{
+        Months = $months
+        Height = (200 * $months.count)
+        Title = "My Calendar"
+        HighlightDate = $HighlightDate
+        Font = $Font
+        FontStyle = $FontStyle
+        FontWeight = $FontWeight
+    }
+
+    Write-Verbose "Using these parameters"
+    $myparams | Out-String | Write-Verbose
+
+    $newRunspace = [RunspaceFactory]::CreateRunspace()
+    $newRunspace.ApartmentState = "STA"
+    $newRunspace.ThreadOptions = "ReuseThread"
+    $newRunspace.Open()
+
+    Write-Verbose "Defining runspace script"
+    $psCmd = [PowerShell]::Create().AddScript( {
+
+        Param (
+            [datetime[]]$HighlightDate,
+            [string]$Font,
+            [string]$FontStyle,
+            [string]$FontWeight,
+            [int]$Height,
+            [string]$Title,
+            [datetime[]]$Months
+        )
+
+        #create a window form.
+        $form = New-Object System.Windows.Window
+
+        $form.AllowsTransparency = $True
+        $form.WindowStyle = "none"
+        #the title won't be shown when window style is set to none
+        $form.Title = $Title
+        $form.Height = $height
+        $form.Width = 200
+
+        $bg = new-object System.Windows.Media.SolidColorBrush
+
+        $form.Background = $bg
+        #color is set for development purposes. It won't be seen normally.
+        $form.Background.Color = "green"
+        $form.background.Opacity = 0
+
+        $form.Add_MouseLeftButtonDown( {$form.DragMove()})
+
+        #add event handlers to adjust opacity by using the +/- keys
+        $form.add_KeyDown( {
+
+                switch ($_.key) {
+                    {'Add', 'OemPlus' -contains $_} {
+                        foreach ($cal in $myCals) {
+                        If ($cal.Opacity -lt 1) {
+                            $cal.Opacity = $cal.opacity + .1
+                            $cal.UpdateLayout()
+                        }
+                    }
+                    }
+                    {'Subtract', 'OemMinus' -contains $_} {
+                        foreach ($cal in $myCals) {
+                        If ($cal.Opacity -gt .2) {
+                            $cal.Opacity = $cal.Opacity - .1
+                            $cal.UpdateLayout()
+                        }
+                      }
+                    }
+                }
+            })
+
+        $stack = $stack = New-object System.Windows.Controls.StackPanel
+        $stack.Width = $form.Width
+        $stack.Height = $form.Height
+        $stack.HorizontalAlignment = "center"
+        $stack.VerticalAlignment = "top"
+
+        #create an array to store calendars so that opacity can be
+        #set for multiple calendars in unison
+        $myCals = @()
+        foreach ($month in $months) {
+            $cal = New-Object System.Windows.Controls.Calendar
+            $cal.DisplayMode = "Calendar"
+
+            $cal.Opacity = 1
+            $cal.FontFamily = $font
+            $cal.FontSize = 24
+            $cal.FontWeight = $FontWeight
+            $cal.FontStyle = $fontStyle
+
+            $cal.DisplayDateStart = $month
+
+            $cal.HorizontalAlignment = "center"
+            $cal.VerticalAlignment = "top"
+
+            $cal.SelectionMode = "multipleRange"
+            if ($highlightdate) {
+                foreach ($d in $HighlightDate) {
+                    if ($d.month -eq $month.Month) {
+                        $cal.SelectedDates.add($d)
+
+                    }
+                }
+            }
+
+            $cal.add_DisplayDateChanged( {
+                    # add the selected days for the currently displayed month
+                    $cal | out-string | write-host
+                    [datetime]$month = $cal.Displaydate
+                    if ($highlightdate) {
+                        foreach ($d in $HighlightDate) {
+                            if ($d.month -eq $month.Month) {
+                                $cal.SelectedDates.add($d)
+                            }
+                        }
+                    }
+                    $cal.UpdateLayout()
+                })
+
+            $stack.addchild($cal)
+            $myCals+=$cal
+        }
+
+        $btn = New-Object System.Windows.Controls.Button
+        $btn.Content = "_Close"
+        $btn.Width = 75
+        $btn.VerticalAlignment = "Bottom"
+        $btn.HorizontalAlignment = "Center"
+        $btn.Opacity = 1
+        $btn.Add_click( {$form.close()})
+
+        $stack.AddChild($btn)
+
+        $form.AddChild($stack)
+        $form.ShowDialog() | out-null
+        })
+
+
+    $pscmd.AddParameters($myparams) | Out-Null
+    $psCmd.Runspace = $newRunspace
+    Write-Verbose "Invoking calendar runspace"
+    $psCmd.BeginInvoke() | Out-Null
+    Write-Verbose "Ending $($myinvocation.mycommand)"
+
+} #close Show-GuiCalendar
+
 
 #a helper function to retrieve names
 function Get-MonthsbyCulture {